과제진행3.1, 3.2
#개인과제 요구사항 3.1, 3.2 수행
3.1 1. 현재 사칙연산 계산기는 +, -, *, /, % 이렇게 총 5가지 연산 타입으로 구성되어있습니다.
    - Enum 타입을 활용하여 연산자 타입에 대한 정보를 관리하고 이를 사칙연산 계산기 ArithmeticCalculator 클래스에 활용 해봅니다.
3.2 2. 지금까지는 ArithmeticCalculator, 즉 사칙연산 계산기는 양의 정수(0 포함)를 매개변수로 전달받아 연산을 수행했습니다.
    - 이제부터는 양의 정수 뿐만 아니라 실수, 즉 double 타입의 값을 전달 받았을 경우에도 연산이 수행되도록, 즉 피연산자를 여러 타입으로 받을 수 있도록 기능을 확장하고 싶습니다.
        - ArithmeticCalculator 클래스의 연산 메서드(`calculate`)
    - 위 요구사항을 만족할 수 있도록 ArithmeticCalculator 클래스를 수정합니다. (제네릭스)
        - 추가적으로 수정이 필요한 다른 클래스나 메서드가 있다면 같이 수정 해주세요.

+++ spartaAssignment/src/AbstractFBOperations.java
@@ -1,3 +1,3 @@
-public interface AbstractFBOperations {
-    public abstract int operate(int num1, int num2);
+public interface AbstractFBOperations{
+    public abstract Double operate(Double num1, Double num2);
 }
